pyproject.toml
[tool.poetry]
name = "bedrock-contextual-retrieval"
version = "0.1.0"
description = "AWS Bedrock based Contextual Retrieval implementation"
authors = ["yunwoong <yunwoong7@gmail.com>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
asgiref = "3.8.1"
backoff = "2.2.1" 
bcrypt = "4.2.1"
boto3 = "1.35.73"
botocore = "1.35.73"
build = "1.2.2.post1"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
chroma-hnswlib = "0.7.6"
chromadb = "0.5.21"
click = "8.1.7"
colorama = "0.4.6"
coloredlogs = "15.0.1"
deprecated = "1.2.15"
durationpy = "0.9"
fastapi = "0.115.5"
filelock = "3.16.1"
flatbuffers = "24.3.25"
fsspec = "2024.10.0"
google-auth = "2.36.0"
googleapis-common-protos = "1.66.0"
grpcio = "1.68.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.28.0"
huggingface-hub = "0.26.3"
humanfriendly = "10.0"
idna = "3.10"
importlib_metadata = "8.5.0"
importlib_resources = "6.4.5"
jmespath = "1.0.1"
kubernetes = "31.0.0"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
mmh3 = "5.0.1"
monotonic = "1.6"
mpmath = "1.3.0"
numpy = "2.1.3"
oauthlib = "3.2.2"
onnxruntime = "1.20.1"
opentelemetry-api = "1.28.2"
opentelemetry-exporter-otlp-proto-common = "1.28.2"
opentelemetry-exporter-otlp-proto-grpc = "1.28.2"
opentelemetry-instrumentation = "0.49b2"
opentelemetry-instrumentation-asgi = "0.49b2"
opentelemetry-instrumentation-fastapi = "0.49b2"
opentelemetry-proto = "1.28.2"
opentelemetry-sdk = "1.28.2"
opentelemetry-semantic-conventions = "0.49b2"
opentelemetry-util-http = "0.49b2"
orjson = "3.10.12"
overrides = "7.7.0"
packaging = "24.2"
posthog = "3.7.4"
protobuf = "5.29.0"
pyasn1 = "0.6.1"
pyasn1_modules = "0.4.1"
pydantic = "2.10.2"
pydantic_core = "2.27.1"
pygments = "2.18.0"
pypdf2 = "3.0.1"
pypika = "0.48.9"
pyproject_hooks = "1.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pyyaml = "6.0.2"
rank-bm25 = "0.2.2"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.9.4"
rsa = "4.9"
s3transfer = "0.10.4"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
starlette = "0.41.3"
sympy = "1.13.3"
tenacity = "9.0.0"
tokenizers = "0.20.3"
tqdm = "4.67.1"
typer = "0.15.0"
typing_extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.32.1"
uvloop = "0.21.0"
watchfiles = "1.0.0"
websocket-client = "1.8.0"
websockets = "14.1"
wrapt = "1.17.0"
zipp = "3.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"